+++ src/services/price/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./price.types";
 export * from "./oracle.types";
+export * from "./multiplier.types";